requirements.txt
ccxt==1.81.50
loguru==0.6.0
pandas==1.4.2
PyMySQL==1.0.2
pytest==7.1.2
python-dotenv==0.20.0
websocket-client==1.3.2
